feat(base): Propose a basic implementation for Coin Coin command
Though it does not add any real usefull content to the bot,
I found the idea if the coin coin command quite funny and
simple enougth to make a good first test for code architecture.

I added a simple use case to start following or unfollowing an user
and sat up a simple SQLite database to keep track of it through
server restart.

This commit does not contain Transport part not test for now.
It mainly aim to discuss organization and propose a way through
it as I don't think it would be a good idea to mix code
organizations in a single repository ;)