refactor: Still aiming to discuss Clean Archi :p
Adapt Grue code to clean archi as I am used to do it.
Correctly splitted code in client between usecase, transport
and modules. Though as Role management is specific to Discord
I choose to not abstract discord entities and logics witch
give an usecase coupled to Discord but I don't think
it should be separated cause discord notion of role management
is very specific to itself.