feat!: Rework code architecture and propose an hello world.
- Current code architecture is too heavy for open source project
and do not let people easily in if they are not used to it.
We decided to refactor it to a simpler base witch can be learned
without pre requisites and let you code using your own habits.

- Enforce some tools to format code
- Add CONTRIBUTING rules and Makefile to ease setup, running && testing

This proposition looks a bit heavy for a simple hello world but it
should stay clean with more complex usecases.

It enforces concern separation as discord part should only filter event
before calling correct module entrypoint, dissociate services from
modules and allow core feature to stay core ^^